Move TestRestart to internal_test
to avoid exposing a private method (restart)
That required two test helpers that
eventually assert on file paths to be moved from
daemon_test into daemontestutils
so we can import in internal_test and daemon_test.
