+++ windows-agent/internal/daemon/daemon_test.go
@@ -3,7 +3,6 @@ package daemon_test
 import (
 	"context"
 	"errors"
-	"io/fs"
 	"net"
 	"os"
 	"path/filepath"
@@ -99,7 +98,7 @@ func TestStartQuit(t *testing.T) {
 					return string(addrContents) != "# Old port file"
 				}, 5*time.Second, 100*time.Millisecond, "Pre-existing address file should be overwritten after dameon.New()")
 			} else {
-				requireWaitPathExists(t, addrPath, "Serve should create an address file")
+				daemontestutils.RequireWaitPathExists(t, addrPath, "Serve should create an address file")
 				addrContents, err = os.ReadFile(addrPath)
 				require.NoError(t, err, "Address file should be readable")
 			}
@@ -155,114 +154,7 @@ func TestStartQuit(t *testing.T) {
 
 			require.NoError(t, <-serveErr, "Serve should return no error when stopped normally")
 			requireCannotDialGRPC(t, address, "No new connection should be allowed when the server is no longer running")
-			requireWaitPathDoesNotExist(t, addrPath, "Address file should have been removed after quitting the server")
-		})
-	}
-}
-
-func TestRestart(t *testing.T) {
-	t.Parallel()
-
-	testsCases := map[string]struct {
-		afterQuit     bool
-		beforeServing bool
-		cancelEarly   bool
-
-		wantAddrFileDeleted bool
-		wantServeErr        bool
-	}{
-		"Success": {},
-		"Does nothing when the context is cancelled":  {cancelEarly: true, wantAddrFileDeleted: true, wantServeErr: true},
-		"Does nothing when daemon is not serving yet": {beforeServing: true},
-		"Does nothing when the daemon is done":        {afterQuit: true, wantAddrFileDeleted: true},
-	}
-
-	for name, tc := range testsCases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-			addrDir := t.TempDir()
-
-			registerer := func(context.Context, bool) *grpc.Server {
-				server := grpc.NewServer()
-				var service testGRPCService
-				grpctestservice.RegisterTestServiceServer(server, service)
-				return server
-			}
-
-			d := daemon.New(ctx, registerer, addrDir)
-
-			serveErr := make(chan error, 1)
-
-			if tc.beforeServing {
-				go func() {
-					d.Restart(ctx)
-					serveErr <- nil
-				}()
-
-				select {
-				case <-time.After(100 * time.Millisecond):
-					require.Fail(t, "Restart should return immediately when daemon is not serving")
-				case <-serveErr:
-					// proceed.
-				}
-			}
-
-			go func() {
-				serveErr <- d.Serve(ctx)
-				close(serveErr)
-			}()
-
-			addrPath := filepath.Join(addrDir, common.ListeningPortFileName)
-
-			var err error
-			requireWaitPathExists(t, addrPath, "Serve should have created a .address file")
-			addrSt, err := os.Stat(addrPath)
-			require.NoError(t, err, "Address file should be readable")
-
-			if tc.afterQuit {
-				d.Quit(ctx, false)
-			}
-			if tc.cancelEarly {
-				cancel()
-			}
-			// Now we know the GRPC server has started serving.
-			d.Restart(ctx)
-
-			// d.Serve() shouldn't have exitted with an error yet at this point.
-			select {
-			case err := <-serveErr:
-				if tc.wantServeErr {
-					require.Error(t, err, "Serve should return with error when stopped by the context")
-				} else {
-					require.NoError(t, err, "Restart should not have caused Serve() to exit with an error")
-				}
-			case <-time.After(100 * time.Millisecond):
-				// proceed.
-			}
-
-			if tc.wantAddrFileDeleted {
-				requireWaitPathDoesNotExist(t, addrPath, "Address file should have been removed after quitting the server")
-				return
-			}
-
-			requireWaitPathExists(t, addrPath, "Restart should have caused creation of another .address file")
-			// Contents could be the same without our control, thus best to check the file time.
-			newAddrSt, err := os.Stat(addrPath)
-			require.NoError(t, err, "Address file should be readable")
-			require.NotEqual(t, addrSt.ModTime(), newAddrSt.ModTime(), "Address file should be overwritten after Restart")
-
-			// Restart a second time
-			d.Restart(ctx)
-			// d.Serve() shouldn't have exitted with an error yet at this point.
-			select {
-			case err := <-serveErr:
-				require.NoError(t, err, "Restart should not have caused Serve() to exit with an error")
-			case <-time.After(100 * time.Millisecond):
-				// proceed.
-			}
+			daemontestutils.RequireWaitPathDoesNotExist(t, addrPath, "Address file should have been removed after quitting the server")
 		})
 	}
 }
@@ -453,7 +345,7 @@ func TestAddingWSLAdapterRestarts(t *testing.T) {
 
 	addrPath := filepath.Join(addrDir, common.ListeningPortFileName)
 
-	requireWaitPathExists(t, addrPath, "Serve should create an address file")
+	daemontestutils.RequireWaitPathExists(t, addrPath, "Serve should create an address file")
 	addrSt, err := os.Stat(addrPath)
 	require.NoError(t, err, "Address file should be readable")
 
@@ -468,7 +360,7 @@ func TestAddingWSLAdapterRestarts(t *testing.T) {
 		// proceed.
 	}
 
-	requireWaitPathExists(t, addrPath, "Restart should have caused creation of another .address file")
+	daemontestutils.RequireWaitPathExists(t, addrPath, "Restart should have caused creation of another .address file")
 	// Contents could be the same without our control, thus best to check the file time.
 	newAddrSt, err := os.Stat(addrPath)
 	require.NoError(t, err, "Address file should be readable")
@@ -521,7 +413,7 @@ func TestQuitBeforeServe(t *testing.T) {
 	}
 
 	d.Quit(ctx, false)
-	requireWaitPathDoesNotExist(t, filepath.Join(addrDir, common.ListeningPortFileName), "Port file should not exist after returning from Serve()")
+	daemontestutils.RequireWaitPathDoesNotExist(t, filepath.Join(addrDir, common.ListeningPortFileName), "Port file should not exist after returning from Serve()")
 }
 
 // grpcPersistentCall will create a persistent GRPC connection to the server.
@@ -578,49 +470,6 @@ func requireCannotDialGRPC(t *testing.T, addr string, msg string) {
 	time.Sleep(300 * time.Millisecond)
 	validStates := []connectivity.State{connectivity.Connecting, connectivity.TransientFailure}
 	require.Contains(t, validStates, conn.GetState(), "unexpected state after dialing. Expected any of %q but got %q", validStates, conn.GetState())
-}
-
-// requireWaitPathExists checks periodically for the existence of a path. If the path
-// does not exist after waiting for the specified timeout, the test fails. This function
-// is blocking.
-func requireWaitPathExists(t *testing.T, path string, msg string) {
-	t.Helper()
-
-	fileExists := func() bool {
-		_, err := os.Lstat(path)
-		if err == nil {
-			return true
-		}
-		require.ErrorIsf(t, err, fs.ErrNotExist, "could not stat path %q. Message: %s", path, msg)
-		return false
-	}
-
-	require.Eventually(t, fileExists, 5*time.Second, 100*time.Millisecond, "%q does not exists: %v", path, msg)
-
-	// Prevent error when accessing the file right after:
-	// 'The process cannot access the file because it is being used by another process'
-	time.Sleep(10 * time.Millisecond)
-}
-
-// requireWaitPathDoesNotExist checks periodically for the existence of a path. If the path
-// does not exist after waiting for the specified timeout, the test fails. This function
-// is blocking.athDoesNotExist checks periodiclly for the existence of a path. If the path
-// does not exist after waiting for the specified timeout, the test fails. This function
-// is blocking.
-func requireWaitPathDoesNotExist(t *testing.T, path string, msg string) {
-	t.Helper()
-
-	var err error
-	fileDoesNotExist := func() bool {
-		_, err = os.Lstat(path)
-		if err == nil {
-			return false
-		}
-		require.ErrorIsf(t, err, fs.ErrNotExist, "could not stat path %q. Message: %s", path, msg)
-		return true
-	}
-
-	require.Eventually(t, fileDoesNotExist, 100*time.Millisecond, time.Millisecond, "%q still exists: %v", path, msg)
 }
 
 // Our mock GRPC service.
